refactor: Move ReorderScopes into Admin module
